requirements.txt
pandas
snowflake-connector-python
tkinter
